move config in shared & rename shared in SDK
