+++ sdk/config/lighterConfig.ts
