+++ sdk/index.ts
@@ -2,3 +2,4 @@ export * from './event'
 export * from './contract'
 export * from './order'
 export * from './fallback'
+export * from './config'
