+++ tasks/lighter-events.ts
@@ -1,7 +1,6 @@
 import {task} from 'hardhat/config'
-import {getAllLighterEvents, lighterEventToString} from '../shared'
+import {getAllLighterEvents, lighterEventToString} from '../sdk'
 
-// npx hardhat getAllLighterEvents --transactionhash '0x72568e681288efbb79189945ce35c22676133e92c6e83e59737fdc5d2c00011f' --network arbgoerli
 task('getAllLighterEvents')
   .addParam('transactionhash')
   .setDescription('get all lighterEvents emitted by OrderBook')
