+++ tsconfig.json
@@ -11,6 +11,6 @@
     "types": ["@nomiclabs/hardhat-ethers", "@nomiclabs/hardhat-waffle"],
     "resolveJsonModule": true
   },
-  "include": ["./test", "./scripts", "./typechain-types", "./reports", "./config"],
+  "include": ["./test", "./typechain-types", "./sdk/config", "./sdk"],
   "files": ["./hardhat.config.ts"]
 }
